Split TDS server rendering
